Add lint and test workflow (CI) (#13)
### Summary
This adds a workflow to run lint (ESLint + Prettier --check) and tests
(with coverage) using GitHub Actions.

### Related Issue
Closes #6
Based on PR #11, so should wait to merge until after that

### Tests & Checks
This PR serves as the test bed to the workflow. I will push commits with
lint issues and/or failing tests so we can confirm success/failure.