Add a temporary mesh source (#16)
Summary:

This PR adds a plane geometry generator that outputs a temporary data source
(MeshSource). The intention is to generate geometry in the mesh layer and make
it available to the renderer. I consider these changes entirely exploratory and
expect most of this code to change once we learn more about efficient data
fetching patterns. The goal of adding this component is to enable rendering of
a single mesh layer.