Create shader programs and refactor (#18)
Summary:
This PR refactors the shader program class. It renames WebGLShaders to
WebGLShaderProgram and simplifies the API -- instead of adding shaders and
linking the program separately, you now supply the vertex and fragment source
codes to the constructor, and the program links automatically.
Additionally, this PR initializes the Vite GLSL plugin and creates a sample
shader program for meshes.