Render a 2D image from an OME-Zarr 5D volumetric time-series (#28)
### Summary
This adds a rudimentary implementation of asynchronously loading then
rendering a 2D slice from an OME-Zarr multi-scale volume. This
implements enough to satisfy #21, but there are plenty of future design
choices needed before this is more generally useful (e.g. how to support
OME-Zarr images with different dtypes).

This adds a dependency on zarrita as a way to expedite the
implementation of the prototype (and to also exercise zarrita as a solid
dependency that we might want to upstream to), but we may consider
implementing our own zarr chunk fetcher in the future.

<img width="1624" alt="Screenshot 2024-09-16 at 4 14 04 PM"
src="https://github.com/user-attachments/assets/d53d04eb-eae6-4b33-9c8b-2725243decc0">

### Related Issue
Closes #21

### Tests & Checks
Manually tested this with the newly added example.

---------

Co-authored-by: Shlomi Nissan <admin@betamark.com>