Move script to new directory with some development utilities (#54)
### Summary
This moves the script defined in #51 into a new more suitable
sub-directory that is now present after merging with `main`.

It also adds a requirements file for very basic tracking of Python
dependencies and ruff + pre-commit for consistent linting and formatting
of Python source files. Then fixes a couple of minor issues found by
ruff.