Initial Line renderable object + sample layer (#46)
### Summary
This adds a new "Line" renderable object, and a sample layer that
renders a single pre-defined line.

This uses [screen-space
projected](https://mattdesl.svbtle.com/drawing-lines-is-hard#screenspace-projected-lines_2)
line rendering method. This PR does not implement any additional end
caps or joins.

Here is what the example looks like now:
![Screenshot 2024-09-27 at 12 15
01 PM](https://github.com/user-attachments/assets/3efb1906-2ef6-44e7-86fe-9a0032438268)

Here's a debug screenshots to give an idea of how this works, from
rendering the same geometry with and LINE_STRIP:
![Screenshot 2024-09-27 at 12 17
46 PM](https://github.com/user-attachments/assets/a25c6b0b-d315-46eb-b802-0dee01e63bce)

This also moves examples into subdirs, and creates a very simple index
that links to each. The `examples` folder is now also watched by the dev
server for reloading.

### Related Issue
Closes #29

### Tests & Checks
So far this has only been tested/debugged manually and visually.

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>