Load video from prototype OME-Zarr volume and render frames based on slider value (#65)
### Summary
This adds a new `ImageSeriesLayer` type that loads multiple images from
the prototype OME-Zarr and allows one to be rendered at a time. It adds
a new example to demonstrate such a usage and also modifies the
prototype React app to do something similar.

I have some more complicated versions of this PR that start to think
about different generalizations that we'll likely need to think about at
some point (e.g. loading m-dimensional image chunks from n-dimensional
image sources), but I kept things as simple as I could in this change.

In the same spirit, there is a lot of opportunity for optimization here.
In particular, support for updating texture data in place or support for
2D texture arrays is an obvious improvement, but likely can be done in
near-term future work.

### Related Issue
Closes #64

### Tests & Checks
I manually tested the new example, the prototype React app, and the
existing examples.

---------

Co-authored-by: Shlomi Nissan <admin@betamark.com>