ProjectedLine fixes (#72)
### Summary
In developing #71 I encountered a problem with the `ProjectedLine`
renderable object. The width of a line from `[-1, 0, 0]` to `[1, 0, 0]`
did not match that of a unit-square (plane).

The problem here was the `offset` in the shader having a homogeneous
component of 1.0. This would cause the line to shrink in screen space.
Removing this also made `currZ` unnecessary - this makes sense to me as
we only want to offset in `x` and/or `y` in screen space.

The other fix here is that the attribute locations for previous/next
positions were reversed. This didn't make much of a difference except
perhaps at the very endpoints of a line.

Finally - I think my pattern of `const previous = path[i - 1] ||
path[i];` should instead be `const previous = path[i - 1] ?? path[i];`
because I was intending this to check for nullish values, non falsey
ones.

### Related Issue
N/A

### Tests & Checks
Tested manually by drawing horizontal and vertical lines on a
unit-square plane with an orthographic camera