Add Layer.transform property and additional AffineTransform features (#79)
### Summary
This adds a `transform` (`AffineTransform`) property to `Layer`.

In the current implementation this transform is composed with the
`RenderableObject.transform` and the camera transform in the `Renderer`
base class, and passed into `renderObject` as `modelView`. I don't think
this is necessarily the right pattern/implementation but hopefully we
can sort that out here.

Other options:
* Pass `layer.transform` directly to `renderObject` (perhaps default to
identity)
* Introduce an `activeLayer` property similar to `activeCamera` on the
renderer, and access the layer transform that way
* Pass the whole layer into `renderObject` along with its object
* Make sure any layer parameters are copied into/onto or passed through
its renderable objects
* Give renderable objects a reference to their "parent" layer

This also adds `scale` and `rotate` method to `AffineTransform`. One
thought after using this a little is that we may also want `setScale`,
`setRotation`, and `setTranslation` methods that don't apply the new
transform incrementally.

Finally, this adds an `inverse` property to `AffineTransform`. I use
this to apply the inverse of the camera transform to the object to get
the `modelView` matrix. This makes more sense to me than using `matrix`
here but I could be thinking about it wrong.

### Related Issue
Closes #73 

### Tests & Checks
Tested manually and created a new example to show how it works.

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>