Load tracks in prototype app (#92)
### Summary
This ports the example from #71 to the Ultrack prototype app. In the
process I also converted the mock data in the app (still hard-coded, now
with 10 "divison" tasks) to match the current format of the backend HITL
server. As part of this process I created some types in the FE to match
the API in the BE.

### Related Issue
Related to #74 (but not sufficient to close)

### Tests & Checks
Tested manually by running `npm run dev -- --mode prototype` and
navigating though tasks, scrubbing time, playing/pausing.



https://github.com/user-attachments/assets/95322c64-893c-4b73-8f43-b28d4ee73717

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>