Allow layer state change callback to be removed (#94)
### Summary
This adds the `Layer.removeStateChangeCallback` to allow a callback to
be removed from those that will be executed on layer state changes.
After adding that I decided to change the name of `Layer.onStateChange`
to `Layer.addStateChangeCallback` to mirror the add/remove event
listener API names.

### Related Issue
Closes #77

### Tests & Checks
I added an automated test to ensure that a callback is not executed
after being added and then removed. I also tested the prototype manually
to ensure that the callback is removed after the initial dev
mount/dismount.