Add pan + zoom for orthographic and perspective cameras (#96)
### Summary
As discussed each camera has two transformation matrices that ultimately
just get composed. This also stores zoom separately and applies it when
updating the projection matrix. For the `OrthographicCamera` it keeps
the frame centered (0, 0) and pans by setting translation on the
`camera.transform`, so I updated it to store just width and height.

I moved the camera controls to their own class in
https://github.com/chanzuckerberg/imaging-active-learning/commit/2f7f7e6a5d579701fed8c4bc00095cdf88136c7d.
This has some oddities as we discussed yesterday, mostly due to
render.canvas being private and the "active" camera switching. I am open
to reverting that (keeping the callbacks in the example for demo
purposes) for to keep the PR a bit less intrusive.

It's an open question how to get the clip-space Z value for the object -
necessary for natural panning with the `PerspectiveCamera`. Right now I
calculate this in the client and just assume the object is at Z = 0 in
world space (and that the camera is looking along Z). Ultimately picking
or some other parameter (i.e. giving a world-space plane) might be a
better solution.

### Related Issue
Related to #89 (but does not close)
Closes #101 

### Tests & Checks
Tested manually by adding pan/zoom mouse controls and a camera toggle to
the `image2d_from_omezarr5d_u8` example.



https://github.com/user-attachments/assets/9a21c836-7039-48e9-9948-1e9b7bee4fbc