[WIP] Connect client to HITL server (#108)
### Summary
This makes the client (prototype frontend) talk to the HITL server
(prototype backend).

This is WIP so feel free to take a look, but I'd like to split it up a
bit to make it easier to review and merge.

What's in here now:
* added `tasks` table to the HITL server DB
* added `answers` table to the HITL server DB
* added an `answer` endpoint to the HITL server
* changed `Answer` in FE to a type that contains a UUID, timestamp, and
sync status
* added effects to sync answers with the server
* icon color in `TaskList` reflects sync status
* control whether using hard-coded data (and mock `answer` endpoint) via
`VITE_MOCK_ULTRACK` env var (allows developing FE without running BE)
* make the HITL server very CORS-permissive (required if FE and BE are
hosted on different origin)
* sort the tracks by ID within a task - this makes every division `red
-> (blue, green)` but I'm not sure it's completely correct (that is, can
a parent have an ID larger than its child?). Maybe we just need to make
sure parent is first.

TODO (will make an issue):
Currently you get a new UUID for each `Answer` when the page loads.
Answers can then be overwritten if you go back to a given task and
re-answer, it will overwrite the previous answer in the DB. If the page
is reloaded or refreshed, you will get new UUIDs and can re-submit _new_
answers. It's up to the server to do what it wants with multiple answers
to one task. An improvement may be adding a concept of a user or session
ID (no need to be secure), such that answers can be changed within the
same session, and perhaps sessions could be resumed

### Related Issues
#105
#106

### Tests & Checks
Manual testing