Add instructions and helper script for running the prototype on Bruno (#117)
### Summary
This adds a helper script and some instructions for running the
prototype app on Bruno (BioHub HPC). This was our stated goal deployment
state for Q4 2024.

Hopefully most of this is pretty clear. I was happy to see no necessary
code changes. I tried to include as much setup as possible in the slurm
script, so it sets up node version manager, node, and uv.

Future improvements include:
* First check for node version instead of blindly installing it
* Run production versions of the servers
* Document process for _users_ of the app, and look into simplifying
this - right now one person needs to launch the slurm job, and others
need to set up the SSH tunnel to connect to the servers.
* Consider splitting jobs for FE + BE servers - FE will be very cheap,
but BE might be co-located with compute and shorter lived. Alternatively
the BE could kick off its own additional slurm jobs (still probably good
to split FE + BE).

### Related Issue
Closes #114 

### Tests & Checks
With this configuration I can interact with the application from my
laptop, and later inspect the SQLite file on Bruno to confirm the
answers have synced properly.

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>