Use bounding rect offset of canvas in client-to-clip transform (#127)
### Summary
This fixes the zoom centering bug described in #124 by using the canvas
element's
[`getClientBoundingRect`](https://developer.mozilla.org/en-US/docs/Web/API/Element/getBoundingClientRect)
in the `Renderer.clientToClip` transform. This accounts for the more
complex layout in the prototype app, which has components to the left of
the canvas.

It also fixes a pan bug in the prototype app by removing controls on the
dismount of the renderer component.

### Related Issue
Closes #124

### Tests & Checks
I manually tested my changes with the tracks example and the prototype.