Support system-dependent theme in prototype (#129)
### Summary
This adds a wrapper `ThemedApp` component to add a system-dependent
dark/light theme. I created a wrapper app based on the [SDS
docs](https://sds.czi.design/009eaf17b/v/0/p/169911-theming/b/36245e)
and moved that to a separate module to satisfy the default requirements
of the [fast refresh React
plugin](https://github.com/ArnaudBarre/eslint-plugin-react-refresh).

I also made the background color of the renderer a mutable attribute.
That's not necessary for this PR as I just use a default of (0, 0, 0, 0)
to make it transparent so that it works well in the application in light
and dark mode. But I think it seems generally useful.

I started making changes to the application to use a non-transparent
background color that matches the light/dark theme, which might be
needed to support certain types of blending. But those set of changes
are much more complicated.


https://github.com/user-attachments/assets/3ea18aec-faa0-4cd2-8e74-977a9a6c70ca

### Related Issue
Closes #52
Closes #100 

### Tests & Checks
I manually tested the prototype in both system dark and light mode after
these changes. I also tested the examples to ensure there were no
regressions there either.