Skip tapering if TaperPower <= 0.0 (#137)
### Summary
The way I implemented line tapering in #107 causes problems with
_non-tapered_ lines. I believe this is because glsl has some unexpected
(to me) undefined behavior in the `pow` function, which I use to taper
the line. Lines are tapered by a cosine envelope from [-pi/2, pi/2],
raised to the `TaperPower`. However, `pow(x, y)` [is undefined if x = 0
and y <=
0](https://registry.khronos.org/OpenGL-Refpages/gl4/html/pow.xhtml),
which would be the case at both ends of a line with `TaperPower` set to
0.

### Related Issue
Closes #134 

### Tests & Checks
Tested manually with prototype and examples.