Adding pixi support (#133)
### Summary

If we decide to support OSX and Windows we will need to switch the build
system from `uv` to `pixi` because some dependencies are mainly on
available on `conda-forge`.

### Tests & Checks

I haven't tested if ultrack is still working with `pixi` I need to
advance with the modeling, and I'll come back to this if we decide to
support other platforms.