Make region coordinates and geometry vertices consistent (#139)
### Summary
This makes the loading input region coordinates and the output geometry
vertex coordinates consistent for image layers. This is demonstrated in
the u16 example, which defines a sub-region in the x-y plane which is
used for both loading and for the orthographic camera.

This also uses OME-Zarr dataset translation values if they are present
and generally simplifies reading the dataset transformation values based
on the restricted specification described in OME-NGFF.

### Related Issue
Closes #35
Closes #38

### Tests & Checks
I tested the development examples.