Add support for loading and rendering float32 data (#148)
### Summary
This adds support for loading a 32-bit floating point image chunk and
using it as a texture. The definitions and logic follow the existing
style and make no attempt to abstract the repetition and redundancy. It
also adds an example that uses a CryoET OME-Zarr tomogram as a data
source.

This PR does does not add support for the equivalent texture array,
mostly because that would need another example and/or other tests.

### Related Issue
Closes #146

### Tests & Checks
I manually tested the new example to check consistency with napari (up
to contrast limits/scaling), and also checked the existing examples
continue to work as expected.