Basic multi-channel support for image layer + float data (#164)
### Summary
This PR adds support for float32 image chunks in the Texture2DArray
class, which is used in the ImageSeriesLayer for multi-channel support.
This also extends that basic multi-channel support to the regular
ImageLayer.

### Related Issues
#146 
#141 

### Tests & Checks
Manual testing of examples and ultrack prototype
Testing the HCS example with data provided by BioHub SF for the
OrganelleBox project