Use complete NGFF types in OME Zarr loader (#157)
### Summary
This uses the newly generated NGFF types in the existing OME-Zarr
loader.

It is an early work in progress, because type information is not being
generated for the referenced parts of the schema (e.g.
coordinateTransformations, axes) and because validation fails with the
uint8 data used for the Ultrack prototype (need to check if that is
actually a data error).

### Related Issue
Closes #145

### Tests & Checks
<!---
  How did you validate that your changes were implemented correctly?
  
  Please think about the following prompts:
  1. I wrote automated tests.
  2. I manually tested my changes, and here's what I did:
-->

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>