Allow contrast limits for multi-channel images (#152)
### Summary
This allows the contrast limits of an image to be configured on layer
creation. These limits are defined by a low and high value per channel,
which the current fragment shader uses to rescale the values.

The contrast limits are passed down to the supporting `Texture` objects
and used by the renderer to set them as a uniform in the related
fragment shaders.

I also changed the image and image-series layer constructors to use
de-structuring, which is not required, but feels more ergonomic in the
long term and helps readability here too.

### Related Issue
Closes #32
Related to #141 

### Tests & Checks
I manually tested the existing image examples and the Ultrack prototype.

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>