Reorganize repo into workspaces (#172)
### Summary
This reorganizes the repo into separate workspaces. This will allow us
to create separate packages for our core repo, as well as a React
component library that depends on the core.

Testing and help updating the README are appreciated!

### Related Issue
Closes #154

### Tests & Checks
I tested I was able to
* build the library with `npm run build` (repo root)
* run the examples with `npm run examples` (repo root)
* run tests in the core workspace
* run the ultrack prototype with `npm run ultrack` (repo root) and `npm
run dev` (ultrack workspace) using `VITE_MOCK_ULTRACK=true`

### Known issues
There are some problems with _types_ being exported properly for our
library such that they help in the other workspaces.
Also this makes it a little harder to auto-reload during development.
Right now I have to `npm run build` the library (and maybe restart the
ultrack dev server) for changes to take up in the ultrack workspace, for
example.