Initial React component library (workspace) (#173)
Merge after #172 

### Summary
This copies the demo component from #155 into a new workspace for react
components. This gives us a place to put React components - e.g. to be
used on the OrganelleBox portal.

This still leaves open actually making/improving those components, and
there may be additional work to package this (and core) workspace for
use in an external project. Hopefully with this merged we can at least
start testing that next week.

### Related issues
#154 
#142 

### Known issues
I'm still having trouble properly using types from `core` in `react`.
The `react` workspace seems aware of the main exported types/classes,
but not for example the type signatures of constructors, functions, etc.