OmeZarrImageViewer component with declarative Renderer (#186)
### Summary
This refactors the `Renderer` component to be more declarative. For now
it's still basically intended for 2D, but that's mostly a limitation of
our camera controls.

This also then creates a new `OmeZarrImageViewer` component that is
intended to be the component used on the organellebox portal. This
component takes a URL for a Zarr and a `Region`. The component handles
creating the source/loader and setting up an `ImageLayer` to be passed
to the renderer. It also attempts to read/set channel information
(color, contrast limits) from the `omero` section of the zattrs for the
image. Finally, it attempts to set the initial camera frame based on the
image size once the region is loaded.

Main changes in `core` are:
* Add `extent` to `ImageLayer`
* Add `metadata` property to `OmeZarrImageLoader` to provide access to
omero channel data
* Some orthographic camera implementation cleanup

### Related Issue
Closes #176
(Integration with #183 is TODO and required for #142 to be closed)

### Tests & Checks
Tested manually using
20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr

---------

Co-authored-by: phoenixAja <plogan@chanzuckerberg.com>