Add optional scale prop to OmeZarrImageViewer (+source +loader) (#204)
### Summary
The organelle box folks have provided us with 4-deep multiscale images.
This provides a way to allow them to specify a scale to display. The
only option they have without this is to re-process their data such that
the lowest scale is the size they want to load in our viewer (that's how
our viewer currently works).

I'm not totally sure I love this implementation, so I'm wide open to
alternatives, but I think this is mostly "good enough" for now. The API
preserves the idea that the default (undefined) option will
intelligently load the proper scale in the future.

This allows indexing with `>=0` to select a specific scale, or `<0` to
select starting at the smallest scale. It will then clamp to the bounds
so it will compensate for lack of a-priori knowledge of the number of
scales and not throw errors. The default behavior remains the same and
will load the _last_ scale (equivalent to passing `-1`, typically the
smallest image / coarsest resolution) for now.

### Tests & Checks
Added a `scale` prop to `OmeZarrImageViewer` and you can select
different scales as described above by changing it.