Require fully-specified region when loading data (#212)
### Summary
This updates the `Region` type to be a tagged union that can be an
interval, point, or "full" (to span an entire dimension of unknown
size). However you now need to specify all dimensions, rather than
implicitly loading the full span of any omitted dimension. This does
require more a priori knowledge of the saved data, but is more explicit.

In practice you generally needed to know this information anyway, and
flexibility for this is better handled at the application layer.

### Related Issue
N/A

### Tests & Checks
Tested all examples, react demo, and ultrack prototype.