ImageSeriesLayer refactor (#217)
### Summary
This replaces #207 and instead of adding a new layer, improves the
`ImageSeriesLayer` to support scaled dimensions.

I also changed the API to allow sync calls to set the slice either by
index or by "position" (scaled units within the specified region) with
`setIndex` and `setPosition`. I prefer this because it replaces the need
to put `setIndex` in a callback that listens for a state change.

This uses a loading token to make sure that the most recent `setIndex`
or `setPosition` call is respected if data needs to be loaded
asynchronously. This enables lazy-loading slices as they are selected.
To mimic the previous behavior you can use `preloadSeries` - this will
still show slices as they are ready instead of waiting for all to load,
which I think is an improvement.

I added a new example for this, but it pretty much replaced the old
CryoET example so I removed that one. The example (especially "Load all
slices") is still _way_ slower than it needs to be without #206 or
something like it.

### Related Issue
Obviates #207 
Merge after #212
([compare](https://github.com/chanzuckerberg/imaging-active-learning/compare/aganders3/full-region...aganders3/image-series-refactor))

### Tests & Checks
Tested all examples, react demo, and ultrack prototype

---------

Co-authored-by: Shlomi Nissan <admin@betamark.com>