OrganelleBox V1 component (#220)
### Summary
Introduces a new "load high res" button based on BioHub SF requests and
@Janeece's designs:
![Screenshot 2025-03-13 at 12 02
19 PM](https://github.com/user-attachments/assets/56adb621-c09d-4add-bc41-7dd6d30ce7ab)

I also added a "reset channels" button to the channel controls.
@phoenixAja I simplified the code in the channel control list to not do
full two-way syncing of the controls with the layer. I think this may be
"good enough" for now but lmk what you think.

I'm still concerned a "reset camera" button would clutter the viewer, so
I left it out. Previous discussions suggested a page refresh may be
sufficient. Otherwise maybe it can be implemented as a button on the
portal.

Here is a recording of this branch:


https://github.com/user-attachments/assets/f028964b-d5e8-49da-9dba-8eede7b6f738

### Related Issue
Replaces #208

### Tests & Checks
Tested in react workspace
Added a button to set a new url to test how the carousel will work