fix: fix bugs and quirks in ImageSeriesLayer loading (#245)
### Summary
This fixes a few bugs/quirks in the loading of the ImageSeriesLayer.
More changes are likely in the near future, but this is still an
improved user experience.

The main feature here is that lazy-loading slices no longer puts the
layer into a "loading" state, which causes flashing to white
(background) when changing slices/timepoints. I also simplified the
example for lazy-loading to focus on the loading states/async returns.

At the same time I fixed:
* hidden async behavior in `setPosition` where `setIndex` was not
awaited
* ambiguous return types from `setPosition` and `setIndex` - now the
caller can determine if a request was cancelled (duplicate or obsolete)
* the HCS example was just broken, due to a change to the OrganelleBox
dataset that removed the initial value

### Related Issue
N/A

### Tests & Checks
Tested all examples, including ultrack and react projects