feat: add Points renderable object (#242)
### Summary
This PR adds support for displaying points with a new renderable object.
The implementation is based on [this tech
spec](https://docs.google.com/document/d/1EE2KP2WdeqSRcOrRXy3bNISeYiME8ts2ixPDeWecZ6I/edit?tab=t.0),
which is still under review.

Points are drawn in a single pass with attributes:
* position
* color
* marker

where _marker_ is an integer that indexes into a sprite array on the
renderable object. Then the fragment shader is able to sample from that
sprite array to get the alpha value for the marker (symbol). Right now
the library includes three basic binary symbols calculated on the CPU:
circles, squares, and triangles.

The added example shows fetching [point annotations (picked particles)
and images from the CryoET Data
Portal](https://cryoetdataportal.czscience.com/runs/17682). This PR does
not add a new layer, but the example includes a layer for demonstration
purposes. This layer adjusts point color/size based on proximity to the
current slice. As mentioned in the spec, these changes could also be
implemented in a more advanced shader at some point (similarly for the
`ProjectedLine` renderable).

If dynamically changing attributes (or even the number of points) is a
common operation, we should consider using non-interleaved vertex
attributes and reusable buffers to improve performance.


https://github.com/user-attachments/assets/58650c20-5f03-42d3-8a0f-52e2ff7ed6c9

Here is roughly the same view in Neuroglancer:


https://github.com/user-attachments/assets/4fe3d93e-0914-4398-bf8f-b6b786a925b8


The current motivating project for this work is the "Try CryLens" demo -
they would like to use our library to:
* show CryoET images from the data portal
* allow users to click near a pre-picked particle to "find similar
particles"
* show these similar particles overlaid on the image, perhaps with some
indication of similarity (i.e. distance in embedding space)

### Related Issue
N/A

### Tests & Checks
Current suite of tests passes
New example appears to be correctly rendering point annotations from the
CryoET data portal